Moved providers to presentation layer.
